src/classify-function/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0
boto3>=1.28.0
azure-communication-email>=1.0.0
